Cut per-memory connection overhead by 2/3rds
Having measured connect overhead at about 1k each, try to follow
EndleSSH's lead a bit by keeping a list of connections and handling them
in a single loop with one shared timer.

This cuts out a lot of overhead - no more timer for the write timeout,
no more per-connection timer for the delay between writes, just a big
linked list of connections, and if attempts to write with EWOULDBLOCK
enough times, time it out.

This appears to cut mmeory use to about 330 bytes per connection, which
is still more than I'd like, but is better than it was.

It also makes it a lot faffier, not sure if it's worth it?